requirements.txt
alabaster==0.7.11
amqp==2.5.2
appnope==0.1.0
asgiref==3.3.1
asn1crypto==0.24.0
astroid==1.6.3
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.16
django-admin-list-filter-dropdown==1.0.2
django-admin-numeric-filter==0.1.5
django-admin-rangefilter==0.5.0
django-appconf==1.0.3
django-celery-beat==1.5.0
django-celery-email==2.0.2
django-celery-results==1.1.2
django-celerybeat-status==0.0.10
django-classy-tags==0.9.0
django-cms==3.7.0
django-cookies-samesite==0.7.0
django-cors-headers==3.0.2
django-debug-panel==0.8.3
django-debug-toolbar==2.1
django-extensions==2.2.6
django-filter==2.2.0
django-formtools==2.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-json-field==0.5.7
django-json-widget==1.1.1
django-ltree==0.4
django-model-utils==4.0.0
django-mptt==0.10.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-sekizai==1.0.0
django-sql-utils==0.4.1
django-timezone-field==3.1
django-treebeard==4.3
django-uuidfield==0.5.0
djangocms-admin-style==1.4.0
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
docutils==0.14
drf-nested-routers==0.91
drf-yasg==1.20.0
et-xmlfile==1.0.1
factory-boy==3.0.1
Faker==4.1.3
fcm-django==0.3.2
flower==0.9.2
future==0.18.2
Genshi==0.7.3
geographiclib==1.50
geopy==2.0.0
httpretty==0.9.7
idna==2.8
imagesize==1.0.0
importlib-metadata==0.23
inflection==0.5.1
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.2.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10
joblib==0.11
kombu==4.6.6
lazy-object-proxy==1.3.1
lxml==4.6.1
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.4
odfpy==1.4.1
openpyxl==3.0.1
packaging==17.1
pandas==0.22.0
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pockets==0.6.2
prompt-toolkit==2.0.10
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.19
pyfcm==1.4.7
Pygments==2.4.2
pygraphviz==1.5
PyJWT==1.7.1
pylint==1.8.4
pylint-django==0.11
pylint-plugin-utils==0.2.6
PyNaCl==1.3.0
pyodbc==4.0.28
pyparsing==2.2.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Shapely==1.7.1
simplegeneric==0.8.1
six==1.13.0
snowballstemmer==1.2.1
Sphinx==1.7.7
sphinxcontrib-napoleon==0.6.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
tablib==3.0.0
text-unidecode==1.3
tornado==5.1
traitlets==4.3.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
wrapt==1.10.11
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
zipp==0.6.0
git+https://devxplorer@bitbucket.org/devxplorer/relatorio.git@89e7e2d37a27d81088ce3d7e40e09682bb39ade0
dadata==0.3.0
tzwhere==3.0.3
django-environ==0.4.5
django-sesame==2.4
pytest==6.2.4
pytest-django==4.4.0
pytest-cov==2.12.1
coverage==5.5
django-polymorphic==3.0.0
